+++ FindMyGolda/app/src/main/java/com/example/findmygolda/database/AlertDatabaseDAO.kt
@@ -3,6 +3,7 @@ package com.example.findmygolda.database
 import androidx.lifecycle.LiveData
 import androidx.room.*
 
+
 @Dao
 interface AlertDatabaseDAO {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
@@ -15,7 +16,7 @@ interface AlertDatabaseDAO {
     fun getAllAlerts() : LiveData<List<Alert>>
 
     @Query("SELECT * from alerts WHERE branchId = :branchId ORDER BY time DESC LIMIT 1")
-    fun getLastAlertOfBranch(branchId: Int): Alert?
+    fun getLastAlertOfBranch(branchId: Long): Alert?
 
     @Update
     fun update(alert: Alert)
